Namespace and modularize nucypher sandbox utilities
