+++ tests/blockchain/eth/entities/test_actors.py
@@ -3,7 +3,8 @@ import pytest
 from constant_sorrow import constants
 
 from nucypher.blockchain.eth.actors import Miner, PolicyAuthor
-from nucypher.utilities.blockchain import token_airdrop
+from nucypher.utilities.sandbox.blockchain import token_airdrop
+from nucypher.utilities.sandbox.constants import DEVELOPMENT_TOKEN_AIRDROP_AMOUNT
 
 
 class TestMiner:
@@ -12,7 +13,7 @@ class TestMiner:
     def miner(self, testerchain, three_agents):
         token_agent, miner_agent, policy_agent = three_agents
         origin, *everybody_else = testerchain.interface.w3.eth.accounts
-        token_airdrop(token_agent, origin=origin, addresses=everybody_else, amount=1000000*constants.M)
+        token_airdrop(token_agent, origin=origin, addresses=everybody_else, amount=DEVELOPMENT_TOKEN_AIRDROP_AMOUNT)
         miner = Miner(miner_agent=miner_agent, checksum_address=everybody_else[0])
         return miner
